requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
lxml==4.5.0
newsapi-python==0.2.6
pyTelegramBotAPI==3.6.7
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
